Move stacked plots tests to it's own pluginSpec to simplify tests
